TerminalGat

Documentation for TerminalGat.

TerminalGat.code

TerminalGat.extractcode

TerminalGat.gat

TerminalGat.gat

TerminalGat.gess

TerminalGat.gess

TerminalGat.gode

TerminalGat.@code

TerminalGat.@gearch

TerminalGat.@gode

TerminalGat.@search

TerminalGat.code — Method

gode([io::IO], f, types)

Print a code giving the location of a generic Function definition.

TerminalGat.extractcode — Method

extractcode(lines::Vector{String})

Extract code that reporensents the definition of a function from lines, scanning lines line by line using Meta.parse.

TerminalGat.gat — Method

gat(filename::AbstractString)

Hightlight text This is a thin wrapper for gat command written in Go.

if filename is a Markdown, the --render-markdown option is added.

Example

julia> using TerminalGat
julia> gat("README.md")
julia> gat("src/TerminalGat.jl")

If your terminal supports Sixel, you can display images in your terminal

julia> using TerminalGat
julia> using Plots; plot(sin); savefig("sin.png")
julia> gat("sin.png")

TerminalGat.gat — Method

gat(md::Markdown.MD)

Example

Hightlight docstrings with monokai theme

julia> gat(@doc sin)

TerminalGat.gess — Method

gess(filename::AbstractString)

gess works something like gat + less:

julia> using TerminalGat
julia> gess("Project.toml")

TerminalGat.gess — Method

gat(md::Markdown.MD)

Example

julia> gess(@doc sin)

TerminalGat.gode — Method

gode(io::IO, f, types)

Print a code giving the location of a generic Function definition with syntax highlighting by gat command.

TerminalGat.@code — Macro

@gode(ex0)

Applied to a function or macro call, it evaluates the arguments to the specified call, and returns code giving the location for the method that would be called for those arguments. It calls out to the code function.

TerminalGat.@gearch — Macro

@gearch f [mod]

It works like methods(f, [mod::Module]) with the Fizzy Finder feature. Then print a highlighted code that gives the method definition of f specified by the user.

TerminalGat.@gode — Macro

@gode(ex0)

Applied to a function or macro call, it evaluates the arguments to the specified call, and returns code giving the location for the method that would be called for those arguments. It calls out to the gode function.

TerminalGat.@search — Macro

@search f [mod]

It works like methods(f, [mod::Module]) with the Fizzy Finder feature. Then print a code that gives the method definition of f specified by the user.